Fixed ESP32 is in slave mode Wire 2.0.0

+++ DFRobot_SHT20.h
@@ -66,7 +66,7 @@ public:
    * @return None
    */
   void initSHT20(void);
-  void initSHT20(byte sdaPin, byte sclPin);
+  void initSHT20(int sdaPin, int sclPin);
 
   /**
    * @fn readHumidity
